% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.funs.R
\name{dd.funs}
\alias{dd.funs}
\alias{dd.rec.logistic}
\alias{dd.rec.Ricker}
\alias{dd.rec.BevertonHolt}
\title{density-dependent recruitment functions}
\usage{
dd.funs(N, r, K)
}
\arguments{
\item{N}{Number of individuals or biomass in time t. Can (should) be a 
vector structure according to group_by.}

\item{r}{Baseline or maximum recruitment rate.}

\item{K}{Carrying capacity. Can be a vector for each stage and can vary by
patches.}
}
\description{
density-dependent recruitment functions
}
\note{
\code{dd.funs} is a generic name for the functions documented.
\cr
If called, \code{dd.funs} returns its own arguments.
}
