% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.growth.general.R
\name{dd.growth.general}
\alias{dd.growth.general}
\title{density dependent growth generalized}
\usage{
dd.growth.general(N, r, K, theta = 1)
}
\arguments{
\item{N}{Number of individuals or biomass in time t. Can (should) be a 
vector structure according to group_by.}

\item{r}{Intrinsic rate of increase or baseline growth rate.}

\item{K}{Carrying capacity. Can be a vector for each stage and can vary by
patches.}

\item{theta}{Shape parameter that modifies the growth form from *t* to *t+1*.
If *theta* = 1, then it takes a logitic shape (Beverton-Holt-like); if > 1, 
it becomes more concave; if < 1, it becomes more convex; and if *theta* = 0, 
then it takes an density-dependent exponential shape (Ricker-like).}
}
\description{
Modifies growth of population in time *t + 1* by the population size in 
time *t*.
}
\note{
This produces very similar results as `dd.growth.logistic`.
}
