% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.rec.BevertonHolt.R
\name{dd.rec.BevertonHolt}
\alias{dd.rec.BevertonHolt}
\title{Density-dependent Beverton-Holt recruitment function}
\usage{
dd.rec.BevertonHolt(N, a, b, theta = NA)
}
\arguments{
\item{N}{Number of individuals or biomass in time t. Can (should) be a 
vector structure according to group_by.}

\item{a}{Density-independent recruitment rate.}

\item{b}{Rate of change with increased density.}

\item{theta}{Degree of depensation (Allee effect).}
}
\description{
Density-dependent Beverton-Holt recruitment function
}
\note{
Equations expressed as recruit per capita.
}
