% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.rec.Ricker.R
\name{dd.rec.Ricker}
\alias{dd.rec.Ricker}
\title{Density-dependent Ricker recruitment function}
\usage{
dd.rec.Ricker(N, a, b, theta = NA)
}
\arguments{
\item{N}{Number of individuals or biomass in time t. Can (should) be a 
vector structure according to group_by.}

\item{a}{Density-independent recruitment rate.}

\item{b}{Rate of change with increased density.}

\item{theta}{Degree of depensation (Allee effect). Default is NA.

##' @note
Equations expressed as recruit per capita.}
}
\description{
Density-dependent Ricker recruitment function
}
