% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.identity.matrix.R
\name{is.identity.matrix}
\alias{is.identity.matrix}
\title{TRUE/FALSE test for identity matrix}
\usage{
is.identity.matrix(mat)
}
\arguments{
\item{mat}{A matrix.}
}
\value{
TRUE/FALSE
}
\description{
TRUE/FALSE test for identity matrix
}
\examples{
M <- matrix(nrow = 3)
is.identity.matrix(M)

}
