% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmm.ddf.params.R
\name{spmm.ddf.params}
\alias{spmm.ddf.params}
\title{define density dependence function parameters}
\usage{
spmm.ddf.params(
  f_type = NA,
  s_type = NA,
  stages = NA,
  P = NA,
  BB = NA,
  MM = NA,
  r = NA,
  K = NA
)
}
\arguments{
\item{f_type}{Either Ricker or Beverton-Holt type density dependence on 
recruitment.}

\item{s_type}{Currently implments only logistic type density dependence on 
survival.}

\item{stages}{Number of stages being modeled.}

\item{P}{Vec-permutation matrix.}

\item{BB}{Block diagonal demographic matrix.}

\item{MM}{Block diagonal movement (dispersal) matrix.}

\item{r}{Baseline recruitment.}

\item{K}{Carrying capacity.}
}
\description{
define density dependence function parameters
}
