% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spmm.readxl.R
\name{spmm.readxl}
\alias{spmm.readxl}
\title{read spreadsheets from Excel Workbook (xlsx) formatted for spmm}
\usage{
spmm.readxl(path, filename)
}
\arguments{
\item{path}{path to xlsx workbook}

\item{filename}{name of xlsx workbook}
}
\value{
Returns a list. Objects can be assigned from the list using the 
unpacking / destructuring assignment "%<-%" from the zeallot package.
(metapopbio currently lists zeallot as a dependency).
}
\description{
Read in spmm worksheet data
}
\note{
Workbook should follow template found in extdata.
}
\references{
Schauberger P, Walker A (2022). openxlsx: Read, Write and Edit xlsx Files.
https://ycphs.github.io/openxlsx/index.html, https://github.com/ycphs/openxlsx.
}
