% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto.spmm.R
\name{auto.spmm}
\alias{auto.spmm}
\title{automatically construct and analyze an spmm}
\usage{
auto.spmm(
  path,
  filename,
  ddf = NA,
  H = NA,
  D = NA,
  plot = FALSE,
  ylabs = NA,
  xlabs = NA
)
}
\arguments{
\item{path}{path to xlsx workbook}

\item{filename}{name of xlsx workbook}

\item{ddf}{Density-dependent function parameters (see `?spmm.ddf.params`)}

\item{H}{Harvest mortality. Currently additive mortality assumed.}

\item{D}{A list of three vectors. The first two, `from` and `to`, identify
where deterrence of movement is made. The third, `d`, contains the proportions
by which movement is deterred. Currently deterrence is assumed equal for all
stages.}

\item{plot}{optional; output plots from `auto.spmm`?}

\item{ylabs}{optional; labels for y-axis labels}

\item{xlabs}{optional; labels for x-axis labels}
}
\description{
This function automates construction and analysis of a spatial 
population matrix model, from construction of the vec-permutation matrix to
plotting. The user provides a `path` and `filename` for a spmm formatted
Excel Workbook (.xlsx). 

Future plans include a "report" option that would include an RMarkdown (.Rmd)
and Word Document (.docx) as outputs. Additionally, I am exploring the ability
to "update" a spmm as a convenience function.
}
\note{
Workbook should follow template found in extdata. Objects can be 
assigned from the list using the unpacking / destructuring assignment "%<-%" 
from the zeallot package. (metapopbio currently lists zeallot as a dependency).
}
\references{
Schauberger P, Walker A (2022). openxlsx: Read, Write and Edit xlsx Files. 
https://ycphs.github.io/openxlsx/index.html, https://github.com/ycphs/openxlsx.
}
