% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd.growth.logistic.R
\name{dd.growth.logistic}
\alias{dd.growth.logistic}
\title{Density dependent logistic growth}
\usage{
dd.growth.logistic(N, r = NULL, B = NULL, K)
}
\arguments{
\item{N}{Number of individuals or biomass in time t. Can (should) be a 
vector structure according to group_by.}

\item{r}{Intrinsic rate of increase or baseline growth rate.}

\item{K}{Carrying capacity. Can be a vector for each stage and can vary by
patches.}
}
\description{
Modifies growth of population in time *t + 1* by the population size in 
time *t*.
}
\note{
This produces very similar results as `dd.growth.exponential`.
}
